feat(df-repr/df-bridge): upgrade to datafusion 43.0.0
plus, regressions in sorts (logical plan already has properties)
regressions in empty relations

+++ optd-sqlplannertest/tests/basic/verbose.planner.sql
@@ -26,7 +26,7 @@ select count(*) from t1;
 /*
 PhysicalStreamAgg
 ├── aggrs:Agg(Count)
-│   └── [ 1(u8) ]
+│   └── [ 1(i64) ]
 ├── groups: []
 ├── cost: {compute=5100,io=1000}
 ├── stat: {row_cnt=1000}
